refactor: move useAuthTypeMetadata and useTokenRefresh into lib/auth/hooks
Both hooks are auth-lifecycle concerns and belong alongside useSessionWatcher
in lib/auth/hooks.ts. The accompanying test moves to lib/auth/hooks.test.tsx.
Old files in src/hooks/ deleted; all callsites updated.
